Rework the preference window
The preferences window receives a makeover. A new general tab is introduced for the general program settings, tap speed, sensitivity and index-ring distance sliders.
The trackpad and mouse tabs remain as-is, while the Characters tab loses it's `Advanced` button. The About tab gets a little redesign and is now also home to the Sparkle updater settings.